fuse: don't drop the dcache when invalidating entry
If the directory is in-use (e.g. as the current working directory of a
process) then calling `d_invalidate` calls `d_drop` which causes the
directory to appear as deleted.

Instead, we only need to drop the `fuse_dentry_settime` to 0, as if a
lookup had returned a 0 attribute cache time.

Signed-off-by: David Scott <dave.scott@docker.com>